Revert "After Deployment"
This reverts commit 970c807e5df032f5a3be8fa375ec51e14159de99.

+++ database/migrations/2018_04_20_092058_create_books_table.php
@@ -20,7 +20,7 @@ class CreateBooksTable extends Migration
             $table->string('bookName');
             $table->string('bookAuthor');
             $table->date('publishedYear');
-            $table->float('bookRank',3,2);
+            $table->integer('bookRank');
             $table->string('attachedFile');
             $table->string('cover_image');
             $table->timestamps();
